# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/procurement/workspaces/{orgId}/quotes": {
      "get": {
        "operationId": "ProcurementController_list",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      },
      "post": {
        "operationId": "ProcurementController_create",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
    "/api/v1/procurement/workspaces/{orgId}/quotes/{quoteId}": {
      "get": {
        "operationId": "ProcurementController_get",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      },
      "delete": {
        "operationId": "ProcurementController_remove",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
    "/api/v1/procurement/workspaces/{orgId}/quotes/{quoteId}/reprice": {
      "post": {
        "operationId": "ProcurementController_reprice",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
    "/api/v1/procurement/workspaces/{orgId}/quotes/{quoteId}/finalize": {
      "post": {
        "operationId": "ProcurementController_finalize",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
    "/api/v1/procurement/workspaces/{orgId}/quotes/{quoteId}/items/{itemId}": {
      "patch": {
        "operationId": "ProcurementController_updateItem",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuoteItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
    "/api/v1/procurement/workspaces/{orgId}/quotes/{quoteId}/csv": {
      "get": {
        "operationId": "ProcurementController_csv",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
    "/api/v1/procurement/workspaces/{orgId}/quotes/{quoteId}/pdf": {
      "get": {
        "operationId": "ProcurementController_pdf",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "procurement"
        ]
      }
    },
...
  },
  "info": {
    "title": "PriceLens API",
    "description": "Cross-platform price comparison API: one canonical product per item, with every store listing matched to it and its price history.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "QuoteItemDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255
          },
          "specs": {
            "type": "string",
            "minLength": 0,
            "maxLength": 500
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          },
          "maxUnitPrice": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "minLength": 0,
            "maxLength": 500
          }
        },
        "required": [
          "query"
        ]
      },
      "CreateQuoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "notes": {
            "type": "string",
            "minLength": 0,
            "maxLength": 1000
          },
          "items": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemDto"
            }
          }
        },
        "required": [
          "title",
          "items"
        ]
      },
      "UpdateQuoteItemDto": {
        "type": "object",
        "properties": {
          "canonicalProductId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          },
          "note": {
            "type": "string",
            "minLength": 0,
            "maxLength": 500
          },
          "maxUnitPrice": {
            "type": "number",
            "minimum": 0
          }
        }
      },
...
    }
  }
}